supporting --initially-drop-socket-file - by default gh-ost will not delete an existing socket file   and thus, will fail running if socket file exists. This is the desired behavior. - The flag --initially-drop-socket-file indicates we take responsibility and wish gh-ost to delete this file on startup

+++ go/logic/server.go
@@ -36,7 +36,7 @@ func (this *Server) BindSocketFile() (err error) {
 	if this.migrationContext.ServeSocketFile == "" {
 		return nil
 	}
-	if base.FileExists(this.migrationContext.ServeSocketFile) {
+	if this.migrationContext.DropServeSocket && base.FileExists(this.migrationContext.ServeSocketFile) {
 		os.Remove(this.migrationContext.ServeSocketFile)
 	}
 	this.unixListener, err = net.Listen("unix", this.migrationContext.ServeSocketFile)
